# findBestMatch() tests the inclusion of default partition name